Tests/LibWeb/Text/input/abspos-relayout-anchor-repaint-restyle.html

<!DOCTYPE html>
<script src="include.js"></script>
<style>
    body { margin: 0; }
    #stage { position: relative; width: 400px; height: 300px; }
    #anchor {
        anchor-name: --btn;
        position: absolute;
        left: 60px;
        top: 40px;
        width: 100px;
        height: 30px;
    }
    #anchored {
        position: absolute;
        position-anchor: --btn;
        left: anchor(left);
        top: anchor(bottom);
        width: 120px;
    }
    #inner { width: 80px; height: 20px; }
</style>
<div id="stage">
    <div id="anchor"></div>
    <div id="anchored">
        <div id="inner"></div>
    </div>
</div>
<script>
    // A restyle rebuilds computed insets from the raw properties, degrading a bare anchor()
    // to auto and wiping the values the last full pass resolved. A repaint-only restyle
    // (color) followed by a change inside the box must take the full layout path, which
    // re-resolves the insets and keeps the box anchored.
    function rectToString(rect) {
        return `${rect.left},${rect.top},${rect.width},${rect.height}`;
    }

    asyncTest(async (done) => {
        document.body.offsetHeight;

        document.getElementById("anchored").style.color = "red";
        const partialBefore = internals.partialLayoutCount();
        const fullBefore = internals.fullLayoutCount();
        document.getElementById("inner").style.height = "60px";
        document.body.offsetHeight;
        const partialDelta = internals.partialLayoutCount() - partialBefore;
        const fullDelta = internals.fullLayoutCount() - fullBefore;

        const anchored = document.getElementById("anchored").getBoundingClientRect();
        println(`layout-passes partial=${partialDelta} full=${fullDelta}`);
        println(`anchored=${rectToString(anchored)}`);
        println(partialDelta === 0 && fullDelta === 1
            && rectToString(anchored) === "60,70,120,60" ? "PASS" : "FAIL");
        done();
    });
</script>
