Tests/LibWeb/Text/expected/abspos-relayout-anchor-repaint-restyle.txt

layout-passes partial=0 full=1
anchored=60,70,120,60
PASS
